+++ extensions/ip-geolocation/src/query-ip-geolocation.tsx
@@ -3,7 +3,7 @@ import { useMemo, useState } from "react";
 import { ActionOpenExtensionPreferences } from "./components/action-open-extension-preferences";
 import { IpEmptyView } from "./components/ip-empty-view";
 import { isEmpty } from "./utils/common-utils";
-import { listIcons } from "./utils/constants";
+import { ipListIcons } from "./utils/constants";
 import { useIpGeolocation } from "./hooks/useIpGeolocation";
 
 interface IpArgument {
@@ -41,12 +41,12 @@ export default function QueryIpGeolocation(props: { arguments: IpArgument }) {
       {ipGeolocation.map((value, index) => (
         <List.Item
           key={index}
-          icon={listIcons[index]}
+          icon={ipListIcons[index]}
           title={value[0]}
           subtitle={value[1]}
           actions={
             <ActionPanel>
-              <Action.CopyToClipboard icon={listIcons[index]} title={`Copy ${value[0]}`} content={value[1]} />
+              <Action.CopyToClipboard icon={ipListIcons[index]} title={`Copy ${value[0]}`} content={value[1]} />
               <Action.CopyToClipboard
                 title={`Copy All Info`}
                 content={JSON.stringify(Object.fromEntries(ipGeolocation), null, 2)}