+++ extensions/ip-geolocation/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix the problem that the Raycast window is automatically closed when the Copy IP command is refreshed in the background.
 - Refactor the code to simplify the logic and improve the performance
 - Remove the time display for the time zone
+- Update icons and styles
 
 ## [Fix Hotkey Error] - 2024-05-27